revert rabbitmqctl related changes
After further troubleshooting, I found out that RabbitMQ passwords are being set
through RPM installation scripts, but they are simply not effective if the
system hostname changes before RabbitMQ is started. Sadly this is almost always
the case with docker containers, so it has to be executed after container is
started.